F2P-111 | Fixed issue with Player Hiscores where some ranks were incorrect if multiple players had the same experience in a skill, by moving compareHiscores function to hiscoresUtils and reusing this function for both the main hiscores and player hiscores, and also filtering the hiscores on the player page before sorting
