+++ src/helpers/hiscores/hiscoresUtils.ts
@@ -1,5 +1,6 @@
 import { HiscoreDataRow } from '@globalTypes/Database/HiscoreDataRow'
 import { HiscoresSortField } from '@globalTypes/Database/HiscoresSortField'
+import { HiscoreType } from '@globalTypes/Hiscores/HiscoreType'
 
 export const getTotalExp = (hiscoreRow: HiscoreDataRow) =>
   hiscoreRow.attackxp +
@@ -30,4 +31,28 @@ export const isNotBaselineExp = (hiscore: HiscoreDataRow, propName: string) => {
   } else {
     return hiscore[propName as keyof HiscoresSortField] > 0
   }
+}
+
+export const compareHiscores = (hiscoreType: HiscoreType, playerOne: HiscoreDataRow, playerTwo: HiscoreDataRow) => {
+  type HiscoreSortKey = keyof HiscoresSortField
+  let fieldName: HiscoreSortKey
+
+  switch (hiscoreType) {
+    case 'Overall':
+      fieldName = 'skill_total'
+      break
+    default:
+      fieldName = `${hiscoreType.toLowerCase()}xp` as HiscoreSortKey
+      break
+  }
+
+  if (playerOne[fieldName] > playerTwo[fieldName]) {
+    return -1
+  }
+
+  if (playerOne[fieldName] < playerTwo[fieldName]) {
+    return 1
+  }
+
+  return 0
 }
