+++ src/hooks/useHiscores.ts
@@ -1,37 +1,12 @@
 import { HiscoreDataRow } from '@globalTypes/Database/HiscoreDataRow'
-import { HiscoresSortField } from '@globalTypes/Database/HiscoresSortField'
 import { HiscoreType } from '@globalTypes/Hiscores/HiscoreType'
-import { isNotBaselineExp } from '@helpers/hiscores/hiscoresUtils'
+import { compareHiscores, isNotBaselineExp } from '@helpers/hiscores/hiscoresUtils'
 import axios from 'axios'
 import { Dispatch, SetStateAction, useEffect, useState } from 'react'
 
 const useHiscores = (hiscoreType: HiscoreType, setIsLoading: Dispatch<SetStateAction<boolean>>) => {
   const [rawHiscores, setRawHiscores] = useState<HiscoreDataRow[] | undefined>(undefined)
   const [hiscores, setHiscores] = useState<HiscoreDataRow[] | undefined>(undefined)
-
-  const compareHiscores = (playerOne: HiscoreDataRow, playerTwo: HiscoreDataRow) => {
-    type HiscoreSortKey = keyof HiscoresSortField
-    let fieldName: HiscoreSortKey
-
-    switch (hiscoreType) {
-      case 'Overall':
-        fieldName = 'skill_total'
-        break
-      default:
-        fieldName = `${hiscoreType.toLowerCase()}xp` as HiscoreSortKey
-        break
-    }
-
-    if (playerOne[fieldName] > playerTwo[fieldName]) {
-      return -1
-    }
-
-    if (playerOne[fieldName] < playerTwo[fieldName]) {
-      return 1
-    }
-
-    return 0
-  }
 
   useEffect(() => {
     setIsLoading(true)
@@ -50,7 +25,7 @@ const useHiscores = (hiscoreType: HiscoreType, setIsLoading: Dispatch<SetStateAc
     const propName = hiscoreType === 'Overall' ? 'skill_total' : `${hiscoreType.toLowerCase()}xp`
     const sortedHiscores = rawHiscores
       ?.filter(hiscoreRow => isNotBaselineExp(hiscoreRow, propName))
-      .sort(compareHiscores)
+      .sort((hiscoreRow1, hiscoreRow2) => compareHiscores(hiscoreType, hiscoreRow1, hiscoreRow2))
 
     setHiscores(sortedHiscores)
     setIsLoading(false)
